updates the build to support both 2.9 and 2.10
project is now constructed differently depending on current scalaVersion.
Only 2.9.2 and 2.10.0-SNAPSHOT are supported, which is a strict superset
of versions that have been supported before.

To compile/stage/whatever for 2.9.2 you don't need to do anything special
(everything stays as before, except the minor changes to script generation).

To compile/stage/whatever for 2.10.0 you need to use special SBT syntax
for explicit provision of scalaVersion, e.g. sbt ++ 2.10.0-SNAPSHOT compile.
Scala itself will be downloaded from sonatype, whereas some of the deps
that had to be rebuilt for 2.10.0-SNAPSHOT will be fetched from SublimeScala
repository (https://bitbucket.org/sublimescala/nexus/raw/master/snapshots/).

Dependency on an external repository is nasty, but it won't last for long,
because the final release of 2.10.0 is going to happen quite soon, plus
I commit to maintaining repo until rebuilt artifacts are deployed to sonatype.
Moreover it's only necessary for the experimental 2.10.0 support, so it isn't
going to affect the normal development process.