SymbolInfo now infers positions from class files
If a symbol is loaded from a class file, its pos will be set to NoPosition.

According to Iulian Dragos this is normal, and a conventional way of dealing
with that is loading the SourceFile attribute from the corresponding class file
and resolving that relative path using editor-specific knowledge about the
project serviced by the presentation compiler.

This is annoying, but what can we do. I'm mostly copying the underlying idea used in
https://github.com/scala-ide/scala-ide/blob/master/org.scala-ide.sdt.core/src/scala/tools/eclipse/LocateSymbol.scala