Take CRLFs into account when internalizing offsets.
The following is only valid for files with CLRF line-endings:
This solution still is far from perfect. The function
`ensime-internalize-offset' will convert offset from CRLF to LF or CR
correctly only if point is currently on the line offset to internalize
corresponds to. This happens because the line-number used for
conversion is calucated from the current location of point.

To sum it up: Expanding selection works correctly as long as :start
and :end are on the same line.